Lint baseline: clippy clean with -D warnings, rustfmt/clippy config, falcon-ocr-eval
- Cargo.toml gains [lints] tables; clippy.toml sets the argument threshold to 8;
  rustfmt.toml sets max_width 120 (the format-only commit follows).
- Fixes for the 24 clippy warnings: is_multiple_of at ten sites, two same-type
  casts, a type alias for the top-K map, non-constant test literals, and the two
  vexp f64 constants rewritten in their shortest round-trip spelling (identical
  bits, checked). Four hot kernels that index by loop variable carry a
  function-level allow with the reason.
- The research binary is renamed falcon-ocr-eval (source file, Cargo target,
  attempt3 scripts, m4_check.sh, which also stops asking for the deleted
  w8-body-kv-bf16 profile).
- tools/check.sh and tools/check.ps1 run fmt, clippy -D warnings, the unit tests
  and optionally the smoke trace hash.

Gates: 104 lib tests; smoke e2dad223...; eval traces 40fd4fb4... / 916b4df8...
unchanged.

+++ src/simd.rs
@@ -204,7 +204,7 @@ impl Simd for Portable {
     #[inline(always)]
     unsafe fn mask_bits(mask: Self::V) -> u32 {
         (0..8).fold(0, |bits, i| {
-            bits | (u32::from(mask[i].to_bits() >> 31) << i)
+            bits | ((mask[i].to_bits() >> 31) << i)
         })
     }
     #[inline(always)]
@@ -903,7 +903,7 @@ mod tests {
     #[test]
     fn native_matches_portable_bitwise() {
         let codes: Vec<i8> = (0..2400)
-            .map(|i| ((i * 37 % 255) as i32 - 127) as i8)
+            .map(|i| (i * 37 % 255 - 127) as i8)
             .collect();
         let scales = values(2400 / 64 + 1, 9)
             .iter()
